renamed task lock to prevent confusion
